pyproject.toml
[project]
name = "webbasedrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.32",
    "langchain-text-splitters>=0.3.10",
    "langserve>=0.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sse-starlette>=3.0.2",
    "streamlit>=1.49.0",
]
